+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Increased action queue check threshold
 - Decreased action registration period
 - Improved default delete policy names and configuration
+- Egress does not generate smoke when artificial-enrichment is disabled
 
 ### Deprecated